+++ common/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/SelectMicrophoneScreen.java
@@ -1,10 +1,8 @@
 package de.maxhenkel.voicechat.gui.audiodevice;
 
+import de.maxhenkel.configbuilder.entry.ConfigEntry;
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
-import de.maxhenkel.voicechat.voice.client.ClientManager;
-import de.maxhenkel.voicechat.voice.client.ClientVoicechat;
-import de.maxhenkel.voicechat.voice.client.SoundManager;
 import de.maxhenkel.voicechat.voice.client.microphone.MicrophoneManager;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.screens.Screen;
@@ -16,9 +14,9 @@ import java.util.List;
 
 public class SelectMicrophoneScreen extends SelectDeviceScreen {
 
-    protected static final ResourceLocation MICROPHONE_ICON = new ResourceLocation(Voicechat.MODID, "textures/icons/microphone.png");
-    protected static final Component TITLE = Component.translatable("gui.voicechat.select_microphone.title");
-    protected static final Component NO_MICROPHONE = Component.translatable("message.voicechat.no_microphone").withStyle(ChatFormatting.GRAY);
+    public static final ResourceLocation MICROPHONE_ICON = new ResourceLocation(Voicechat.MODID, "textures/icons/microphone.png");
+    public static final Component TITLE = Component.translatable("gui.voicechat.select_microphone.title");
+    public static final Component NO_MICROPHONE = Component.translatable("message.voicechat.no_microphone").withStyle(ChatFormatting.GRAY);
 
     public SelectMicrophoneScreen(@Nullable Screen parent) {
         super(TITLE, parent);
@@ -30,12 +28,7 @@ public class SelectMicrophoneScreen extends SelectDeviceScreen {
     }
 
     @Override
-    public String getSelectedDevice() {
-        return VoicechatClient.CLIENT_CONFIG.microphone.get();
-    }
-
-    @Override
-    public ResourceLocation getIcon(String device) {
+    public ResourceLocation getIcon() {
         return MICROPHONE_ICON;
     }
 
@@ -45,16 +38,7 @@ public class SelectMicrophoneScreen extends SelectDeviceScreen {
     }
 
     @Override
-    public String getVisibleName(String device) {
-        return SoundManager.cleanDeviceName(device);
-    }
-
-    @Override
-    public void onSelect(String device) {
-        VoicechatClient.CLIENT_CONFIG.microphone.set(device).save();
-        ClientVoicechat client = ClientManager.getClient();
-        if (client != null) {
-            client.reloadAudio();
-        }
+    public ConfigEntry<String> getConfigEntry() {
+        return VoicechatClient.CLIENT_CONFIG.microphone;
     }
 }